# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PIX Hub API",
        "description": "API para integração com subadquirentes de pagamento (PIX e Saques)",
        "contact": {
            "email": "dev@pixhub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/withdraw": {
            "get": {
                "tags": [
                    "Withdraw"
                ],
                "summary": "Listar saques",
                "description": "Lista todos os saques do usuário autenticado",
                "operationId": "listWithdraw",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de registros",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de saques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "external_id": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Withdraw"
                ],
                "summary": "Criar um novo saque",
                "description": "Cria uma nova solicitação de saque via PIX ou dados bancários",
                "operationId": "createWithdraw",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Valor do saque",
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "pix_key": {
                                        "description": "Chave PIX (CPF, email, telefone ou aleatória)",
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "pix_key_type": {
                                        "description": "Tipo da chave PIX",
                                        "type": "string",
                                        "enum": [
                                            "cpf",
                                            "cnpj",
                                            "email",
                                            "phone",
                                            "random"
                                        ],
                                        "example": "cpf"
                                    },
                                    "bank_code": {
                                        "description": "Código do banco",
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "agency": {
                                        "description": "Agência",
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "account": {
                                        "description": "Número da conta",
                                        "type": "string",
                                        "example": "123456-7"
                                    },
                                    "account_type": {
                                        "description": "Tipo da conta",
                                        "type": "string",
                                        "enum": [
                                            "checking",
                                            "savings"
                                        ],
                                        "example": "checking"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Saque criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Saque criado com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "WD_xyz789"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 500
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "pix_key": {
                                                    "type": "string",
                                                    "example": "12345678901"
                                                },
                                                "pix_key_type": {
                                                    "type": "string",
                                                    "example": "cpf"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2025-11-19T04:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo amount é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Muitas requisições (Rate Limit: 30 req/min)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too Many Attempts."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/withdraw/{id}": {
            "get": {
                "tags": [
                    "Withdraw"
                ],
                "summary": "Obter detalhes de um saque",
                "description": "Retorna os detalhes de uma solicitação de saque específica",
                "operationId": "getWithdraw",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do saque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do saque",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "WD_xyz789"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 500
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "pix_key": {
                                                    "type": "string"
                                                },
                                                "pix_key_type": {
                                                    "type": "string"
                                                },
                                                "completed_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Saque não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Saque não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acesso negado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação via Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}